+++ src/arpack/dnaupd.f
@@ -93,7 +93,7 @@ c     5 below.
 c     IDO = 99: done
 c     -------------------------------------------------------------
 c
-c     BMAT    Character*1.  (INPUT)
+c     BMAT    Character(len=1).  (INPUT)
 c     BMAT specifies the type of the matrix B that defines the
 c     semi-inner product for the operator OP.
 c     BMAT = 'I' -> standard eigenvalue problem A*x = lambda*x
@@ -102,7 +102,7 @@ c
 c     N       Integer.  (INPUT)
 c     Dimension of the eigenproblem.
 c
-c     WHICH   Character*2.  (INPUT)
+c     WHICH   Character(len=2).  (INPUT)
 c     'LM' -> want the NEV eigenvalues of largest magnitude.
 c     'SM' -> want the NEV eigenvalues of smallest magnitude.
 c     'LR' -> want the NEV eigenvalues of largest real part.